// client/src/graphql/generated/QueryPlatforms.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: QueryPlatforms
// ====================================================

export interface QueryPlatforms_platforms {
  __typename: "Platform";
  id: string;
  name: string;
  slug: string;
}

export interface QueryPlatforms {
  platforms: QueryPlatforms_platforms[];
}